API_KERAS/requirements.txt
numpy==1.19.5
Flask==1.1.2
opencv-python==4.5.2.54
imageio==2.9.0
Pillow==8.2.0
imutils==0.5.4
uwsgi
tensorflow==2.5.0
keras
sklearn
